fix: Adjust directory structure
