+++ package.json
@@ -39,6 +39,7 @@
 		"@rollup/plugin-node-resolve": "^13.3.0",
 		"@rollup/plugin-replace": "^2.2.1",
 		"@rollup/plugin-typescript": "^8.5.0",
+		"@rrmc/antd-breadcrumbs": "^1.2.0-beta.0",
 		"@testing-library/jest-dom": "^5.16.4",
 		"@testing-library/react": "13.4.0",
 		"@testing-library/react-hooks": "^8.0.0",
@@ -52,7 +53,7 @@
 		"@typescript-eslint/eslint-plugin": "^4.28.3",
 		"@typescript-eslint/parser": "^4.28.3",
 		"@vitejs/plugin-react": "^3.0.1",
-		"antd": "^5.1.2",
+		"antd": "^5.1.4",
 		"babel-eslint": "^10.1.0",
 		"babel-plugin-dev-expression": "^0.2.2",
 		"babel-plugin-transform-remove-debugger": "^6.9.4",
@@ -78,7 +79,7 @@
 		"prompt-confirm": "^2.0.4",
 		"react": "^18.0.2",
 		"react-dom": "^18.0.2",
-		"react-router-manage": "^1.1.6",
+		"react-router-manage": "^1.2.0-beta.0",
 		"react-test-renderer": "^17.0.2",
 		"rimraf": "^3.0.2",
 		"rollup": "^2.75.5",
